back_end/requirements.txt
fastapi==0.104.1
Pillow==10.1.0
python-multipart==0.0.6
uvicorn==0.24.0.post1
